# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Backend Base NestJS
  version: '1.0'
  description: Documentación Backend base - NestJS
  contact:
    name: agetic@agetic.gob.bo
servers:
  - url: 'https://base.agcs.agetic.gob.bo/ws/api'
paths:
...
  /autorizacion/politicas:
    post:
      operationId: AuthorizationController_crearPolitica
      summary: API para crear una nueva política
      parameters: []
      requestBody:
        required: true
        description: Crear nueva política
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoliticaDto'
      responses:
        '201':
          description: ''
      tags: &ref_0
        - Autorización
      security: &ref_1
        - bearer: []
    patch:
      operationId: AuthorizationController_actualizarPolitica
      summary: API para actualizar una nueva política
      parameters:
        - name: sujeto
          required: true
          in: query
          example: ADMINISTRADOR
          schema:
            type: string
        - name: objeto
          required: true
          in: query
          example: /admin/parametros
          schema:
            type: string
        - name: accion
          required: true
          in: query
          example: create|update|delete|read
          schema:
            type: string
        - name: app
          required: true
          in: query
          example: frontend
          schema:
            type: string
        - name: id
          required: true
          in: query
          schema: {}
      requestBody:
        required: true
        description: Actualizar política
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PoliticaDto'
      responses:
        '200':
          description: ''
      tags: *ref_0
      security: *ref_1
    get:
      operationId: AuthorizationController_listarPoliticas
      summary: API para obtener el listado de politicas
      parameters:
        - name: limite
          required: false
          in: query
          schema:
            minimum: 10
            maximum: 50
            default: 10
            type: number
        - name: pagina
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: filtro
          required: false
          in: query
          schema:
            type: string
        - name: ordenRaw
          required: false
          in: query
          schema:
            type: string
        - required: false
          name: limite
          in: query
          schema:
            minimum: 10
            maximum: 50
            default: 10
            type: number
        - required: false
          name: pagina
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - required: false
          name: filtro
          in: query
          schema:
            type: string
        - required: false
          name: ordenRaw
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_0
      security: *ref_1
    delete:
      operationId: AuthorizationController_eliminarPolitica
      summary: API para eliminar una política
      parameters:
        - name: sujeto
          required: true
          in: query
          example: ADMINISTRADOR
          schema:
            type: string
        - name: objeto
          required: true
          in: query
          example: /admin/parametros
          schema:
            type: string
        - name: accion
          required: true
          in: query
          example: create|update|delete|read
          schema:
            type: string
        - name: app
          required: true
          in: query
          example: frontend
          schema:
            type: string
        - name: id
          required: true
          in: query
          schema: {}
      responses:
        '200':
          description: ''
      tags: *ref_0
      security: *ref_1
  /autorizacion/permisos:
    get:
      operationId: AuthorizationController_obtenerRoles
      summary: API para obtener las politicas definidas en formato CASBIN
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_0
      security: *ref_1
...
  /auth:
    post:
      operationId: AuthenticationController_login
      summary: API para autenticación con usuario y contraseña
      parameters: []
      requestBody:
        required: true
        description: Autenticación de usuarios
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthDto'
      responses:
        '201':
          description: ''
      tags: &ref_6
        - Autenticación
  /cambiarRol:
    patch:
      operationId: AuthenticationController_changeRol
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CambioRolDto'
      responses:
        '200':
          description: ''
      tags: *ref_6
      security:
        - bearer: []
...
  /logout:
    get:
      operationId: AuthenticationController_logoutCiudadania
      summary: API para logout digital
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_6
      security:
        - bearer: []
  /token:
    post:
      operationId: RefreshTokensController_getAccessToken
      summary: API que actualiza el token de acceso usando un refresh token
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenDto'
      responses:
        '201':
          description: ''
      tags: &ref_7
        - Refresh Token
  /{id}:
    delete:
      operationId: RefreshTokensController_eliminarRefreshToken
      summary: API que elimina el refresh token
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_7
      security:
        - bearer: []
...
  /parametros:
    get:
      operationId: ParametroController_listar
      summary: API para obtener el listado de parámetros
      parameters:
        - name: limite
          required: false
          in: query
          schema:
            minimum: 10
            maximum: 50
            default: 10
            type: number
        - name: pagina
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: filtro
          required: false
          in: query
          schema:
            type: string
        - name: ordenRaw
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: &ref_9
        - Parámetros
      security: &ref_10
        - bearer: []
    post:
      operationId: ParametroController_crear
      summary: API para crear un nuevo parámetro
      parameters: []
      requestBody:
        required: true
        description: new Parametro
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearParametroDto'
      responses:
        '201':
          description: ''
      tags: *ref_9
      security: *ref_10
  /parametros/{grupo}/listado:
    get:
      operationId: ParametroController_listarPorGrupo
      summary: API para obtener el listado de parámetros por grupo
      parameters:
        - name: grupo
          required: true
          in: path
          example: TD
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_9
      security: *ref_10
  /parametros/{id}:
    patch:
      operationId: ParametroController_actualizar
      summary: API para actualizar un parámetro
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      requestBody:
        required: true
        description: new Rol
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarParametroDto'
      responses:
        '200':
          description: ''
      tags: *ref_9
      security: *ref_10
  /parametros/{id}/activacion:
    patch:
      operationId: ParametroController_activar
      summary: API para activar un parámetro
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_9
      security: *ref_10
  /parametros/{id}/inactivacion:
    patch:
      operationId: ParametroController_inactivar
      summary: API para inactivar un parámetro
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_9
      security: *ref_10
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    PoliticaDto:
      type: object
      properties:
        sujeto:
          type: string
          example: ADMINISTRADOR
        objeto:
          type: string
          example: /admin/parametros
        accion:
          type: string
          example: create|update|delete|read
        app:
          type: string
          example: frontend
      required:
        - sujeto
        - objeto
        - accion
        - app
...
    AuthDto:
      type: object
      properties:
        usuario:
          type: string
          example: ADMINISTRADOR
          description: Usuario
        contrasena:
          type: string
          example: MTIz
          description: Contraseña
      required:
        - usuario
        - contrasena
    CambioRolDto:
      type: object
      properties:
        idRol:
          type: string
          example: '1'
      required:
        - idRol
    TokenDto:
      type: object
      properties:
        token:
          type: string
          example: ''
      required:
        - token
...
    CrearParametroDto:
      type: object
      properties:
        codigo:
          type: string
          example: TD-CI
        nombre:
          type: string
          example: Cédula de identidad
        grupo:
          type: string
          example: CD
        descripcion:
          type: string
          example: Cédula de identidad
        estado:
          type: string
          example: ACTIVO
      required:
        - codigo
        - nombre
        - grupo
        - descripcion
        - estado
    ActualizarParametroDto:
      type: object
      properties:
        codigo:
          type: string
          example: TD-CI
        nombre:
          type: string
          example: Cédula de identidad
        grupo:
          type: string
          example: CD
        descripcion:
          type: string
          example: Cédula de identidad
        estado:
          type: string
          example: ACTIVO
      required:
        - codigo
        - nombre
        - grupo
        - descripcion
        - estado